04_rag_queue/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
openai==1.106.1
redis==5.0.1
rq==1.15.1
python-multipart==0.0.20
